Switch from use to require to resolve -c check
Due to each module using the other (OBML and OBML::HTML), a syntax check
with perl -c to the OBML module would throw errors, even though these do
not show up in other contexts (like via contracted use or in tests).

To resolve, the use for each is changed to require.

See also #5.